// web/src/gen/model/notificationDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { JsonDocument } from "./jsonDocument";
import type { NotificationSeverity } from "./notificationSeverity";

export interface NotificationDto {
  id: string;
  title: string;
  message: string;
  type: string;
  severity: NotificationSeverity;
  data: JsonDocument;
  isRead: boolean;
  createdAt: string;
}